requirements.txt
Flask
Werkzeug
Pillow
requests
python-dotenv
